Moved correlation to ppl
