--- app/index.ts
@@ -1,2 +1,0 @@
-import {HELLO} from "ppl/javascript"
-console.log(HELLO)
