+++ ppl/correlation/components/ZeekCorrelation.tsx
@@ -3,12 +3,12 @@ import classNames from "classnames"
 import * as d3 from "d3"
 import isEqual from "lodash/isEqual"
 import {zng} from "zealot"
-import brim from "../../brim"
-import {createZeekLog} from "../../brim/zeekLog"
-import {submitSearch} from "../../flows/submitSearch/mod"
-import {viewLogDetail} from "../../flows/viewLogDetail"
-import {UID_CORRELATION_LIMIT} from "../../searches/programs"
-import SearchBar from "../../state/SearchBar"
+import brim from "src/js/brim"
+import {createZeekLog} from "src/js/brim/zeekLog"
+import {submitSearch} from "src/js/flows/submitSearch/mod"
+import {viewLogDetail} from "src/js/flows/viewLogDetail"
+import {UID_CORRELATION_LIMIT} from "src/js/searches/programs"
+import SearchBar from "src/js/state/SearchBar"
 import {BrimEvent} from "../models/BrimEvent"
 import {sort} from "../util/sort"
 import {useDispatch} from "react-redux"
