+++ ppl/correlation/flows/stubs.ts
