+++ ppl/correlation/models/SuricataEvent.ts
