+++ scripts/clean.js
@@ -1,3 +1,4 @@
 let fs = require("fs-extra")
 
 fs.removeSync("dist")
+fs.removeSync("zealot/dist")
