+++ src/js/flows/viewLogDetail.ts
@@ -2,7 +2,7 @@ import LogDetails from "../state/LogDetails"
 import {Thunk} from "../state/types"
 import {zng} from "zealot"
 import {isEqual} from "lodash"
-import {fetchCorrelation} from "../correlation/flows/fetch"
+import {fetchCorrelation} from "ppl/correlation/flows/fetch"
 import Notice from "../state/Notice"
 import ErrorFactory from "../models/ErrorFactory"
 import Current from "../state/Current"
